Low memory dense (#618)
* updated dense kronmult to work directly with the element table

+++ src/device/asgard_kronmult_common.hpp
@@ -33,10 +33,11 @@ namespace asgard::kronmult
  * \param work_per_block is the work that a single thread block will execute
  * \param max_blocks is the maximum number of blocks
  */
-inline int blocks(int work_size, int work_per_block, int max_blocks)
+inline int blocks(int64_t work_size, int work_per_block, int max_blocks)
 {
-  return std::min(max_blocks,
-                  (work_size + work_per_block - 1) / work_per_block);
+  return std::min(
+      max_blocks,
+      static_cast<int>((work_size + work_per_block - 1) / work_per_block));
 }
 
 /*!
